code review

+++ libs/src/classes/crypt/kk_crypt.cpp
@@ -236,14 +236,13 @@ QByteArray KKCrypt::decryptToByteArray(const QByteArray& cypher)
 QString KKCrypt::random_psw(QString s){
 
     int len=20;
-    //    QString s;
     static const char alphanum[] =
             "0123456789"
             "ABCDEFGHIJKLMNOPQRSTUVWXYZ"
             "abcdefghijklmnopqrstuvwxyz";
 
     for (int i = 0; i < len; ++i) {
-        s[i] = alphanum[rand() % (sizeof(alphanum) - 1)];
+        s[i] = alphanum[static_cast<unsigned long>(rand()) % (sizeof(alphanum) - 1)];
     }
